refacto part le sensei du code.
