+++ projects/reactive-forms/src/app/components/form-user/form-user.component.ts
@@ -1,5 +1,5 @@
-import {ChangeDetectionStrategy, Component, inject} from '@angular/core';
-import {FormArray, FormControl, FormGroup, FormsModule, ReactiveFormsModule, Validators} from "@angular/forms";
+import {ChangeDetectionStrategy, Component} from '@angular/core';
+import {FormsModule, ReactiveFormsModule} from "@angular/forms";
 import {MatError, MatFormField, MatHint, MatLabel, MatSuffix} from "@angular/material/form-field";
 import {MatInput} from "@angular/material/input";
 import {AsyncPipe, DatePipe, JsonPipe} from "@angular/common";
@@ -10,8 +10,8 @@ import {MatListOption, MatSelectionList} from "@angular/material/list";
 import {MatIcon} from "@angular/material/icon";
 import {MatDivider} from "@angular/material/divider";
 import {MatDatepicker, MatDatepickerInput, MatDatepickerToggle} from "@angular/material/datepicker";
-import {FormUserService} from "./form-user.service";
-import {passwordMatchValidator, PasswordStateMatcher} from "./form-user.validators";
+import {UserForm} from "./model/user-form";
+import {PasswordStateMatcher} from "./password-state-matcher";
 import {takeUntilDestroyed} from "@angular/core/rxjs-interop";
 import {MatRadioButton, MatRadioGroup} from "@angular/material/radio";
 
@@ -54,75 +54,61 @@ import {MatRadioButton, MatRadioGroup} from "@angular/material/radio";
 })
 export class FormUser {
 
-    formUserService = inject(FormUserService);
+    /** Form */
+    readonly registerForm = UserForm.buildUserForm();
 
+    /** Sub Form Getters */
+    readonly userForm = this.registerForm.controls.user;
+    readonly addressForm = this.registerForm.controls.address;
+    readonly optionalForm = this.registerForm.controls.optional;
+    readonly hobbiesForm = this.optionalForm.controls.hobbiesItems;
 
-    registerForm = new FormGroup(
-        {
-            user: new FormGroup({
-                name: new FormControl('', [Validators.required, Validators.minLength(2), Validators.maxLength(50)]),
-                lastName: new FormControl('', [Validators.required, Validators.minLength(2), Validators.maxLength(50)]),
-                email: new FormControl('', [Validators.required, Validators.email]),
-                age: new FormControl('', [Validators.required, Validators.min(18), Validators.maxLength(100), Validators.pattern('[0-9]')]),
-                phoneNumber: new FormControl('', [Validators.required, Validators.pattern('^[0-9]{10}$')]),
-                password: new FormControl('', [Validators.required, Validators.minLength(8)]),
-                passwordConfirm: new FormControl('', [Validators.required]),
-            }, [passwordMatchValidator]),
-            address: new FormGroup({
-                street: new FormControl('', [Validators.required]),
-                city: new FormControl('', [Validators.required]),
-                zipcode: new FormControl('', [Validators.required, Validators.pattern('^[0-9]{5}$')]),
-            }),
-            optional: new FormGroup({
-                newsletterSub: new FormControl(''),
-                newsletterFrequency: new FormControl('', Validators.required),
-                website: new FormControl('', Validators.pattern('https?://.+')),
-                birthDate: new FormControl('', Validators.required, [this.formUserService.checkBirthDateValidity]),
-                gender: new FormControl('', Validators.required),
-                commentary: new FormControl(''),
-                hobbiesItems: new FormArray<FormControl>([
-                    new FormControl('', [Validators.minLength(3)])
-                ], [Validators.maxLength(4)])
-            }),
-        }
-    );
+    /** Password Form Error Manager */
+    readonly passwordStateMatcher = new PasswordStateMatcher();
 
-    userForm = this.registerForm.controls.user;
-    addressForm = this.registerForm.controls.address;
-    optionalForm = this.registerForm.controls.optional;
-    hobbiesForm = this.optionalForm.controls.hobbiesItems;
-    today = new Date().getDate()
-    passwordStateMatcher = new PasswordStateMatcher();
+    /** Constants */
+    private readonly MAX_HOBBIES = 5;
 
     constructor() {
-        this.optionalForm.controls.newsletterSub.valueChanges.pipe(
-            takeUntilDestroyed()
-        ).subscribe(
-            isSub => {
-                this.optionalForm.controls.newsletterFrequency.reset()
-            }
-        )
+        this.manageNewsletterState();
     }
 
-    createHobbyCtrl(value = '') {
-        return new FormControl(value, [Validators.required, Validators.minLength(3)])
-    }
-
-    addHobby() {
-        const b = this.hobbiesForm.hasError('maxlength');
-        if (!b) {
-            this.hobbiesForm.push(this.createHobbyCtrl())
+    /**
+     * Permet d'ajouter un nouveau hobby au clique sur le bouton d'ajout
+     */
+    addHobby(): void {
+        if (this.hobbiesForm.length <= this.MAX_HOBBIES) {
+            this.hobbiesForm.push(UserForm.buildHobbyCtrl())
         }
     }
 
-    removeHobby(index: number) {
+    /**
+     * Permet de supprimer un hobby au clique sur le bouton de suppression
+     * @param index Position de l'élément à supprimer
+     */
+    removeHobby(index: number): void {
         this.hobbiesForm.removeAt(index);
     }
 
-    submitAuth() {
+    /**
+     * Envoie le formulaire au back-end
+     */
+    submitRegisterForm(): void {
         this.registerForm.markAllAsTouched();
         if (this.registerForm.valid) {
             console.info(this.registerForm.getRawValue())
         }
+    }
+
+    /**
+     * Gère l'état du control "Newsletter" en fonction de la checkbox
+     * @private
+     */
+    private manageNewsletterState(): void {
+        this.optionalForm.controls.newsletterSub.valueChanges.pipe(takeUntilDestroyed())
+            .subscribe(() => {
+                this.optionalForm.controls.newsletterFrequency.reset();
+            }
+        )
     }
 }
