+++ projects/reactive-forms/src/app/components/form-user/form-user.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { FormUserService } from './form-user.service';
+import { FormUserService } from './model/user-form';
 
 describe('FormUserService', () => {
   let service: FormUserService;
