+++ projects/reactive-forms/src/app/components/form-user/model/adresse.model.ts
@@ -1,0 +1,5 @@
+export interface Address {
+    street: string,
+    city: string,
+    zipCode: number,
+}
