+++ projects/reactive-forms/src/app/components/form-user/model/user.model.ts
@@ -1,0 +1,16 @@
+export interface User {
+    name: string,
+    lastname: string,
+    email: string,
+    age: number,
+    birthDate: Date,
+    gender: Gender,
+    phoneNumber: string,
+    password: string,
+}
+
+enum Gender {
+    man = 'Men',
+    women = 'Women',
+    other = 'Other'
+}
